+++ packages/go_router/lib/src/redirection.dart
@@ -6,7 +6,6 @@ import 'configuration.dart';
 import 'logging.dart';
 import 'match.dart';
 import 'matching.dart';
-import 'typedefs.dart';
 
 /// A GoRouter redirector function.
 // TODO(johnpryan): make redirector async
@@ -20,15 +19,12 @@ typedef RouteRedirector = RouteMatchList Function(RouteMatchList matches,
 RouteMatchList redirect(RouteMatchList prevMatchList,
     RouteConfiguration configuration, RouteMatcher matcher,
     {Object? extra}) {
-  RouteMatchList matches;
-
   // Store each redirect to detect loops
   final List<RouteMatchList> redirects = <RouteMatchList>[prevMatchList];
 
   // Keep looping until redirecting is done
   while (true) {
     final RouteMatchList currentMatches = redirects.last;
-
     // Check for top-level redirect
     final Uri uri = currentMatches.location;
     final String? topRedirectLocation = configuration.topRedirect(
@@ -52,11 +48,8 @@ RouteMatchList redirect(RouteMatchList prevMatchList,
       continue;
     }
 
-    // If there's no top-level redirect, keep the matches the same as before.
-    matches = currentMatches;
-
     // Merge new params to keep params from previously matched paths, e.g.
-    // /users/:userId/book/:bookId provides userId and bookId to book/:bookId
+    // /users/:userId/book/:bookId provides userId and bookId to bookgit /:bookId
     Map<String, String> previouslyMatchedParams = <String, String>{};
     for (final RouteMatch match in currentMatches.matches) {
       assert(
@@ -67,46 +60,39 @@ RouteMatchList redirect(RouteMatchList prevMatchList,
       previouslyMatchedParams = match.encodedParams;
     }
 
-    // check top route for redirect
-    final RouteMatch? top = matches.isNotEmpty ? matches.last : null;
-    if (top == null) {
-      break;
+    // check route redirection.
+    RouteMatchList? newMatchList;
+    for (final RouteMatch match in currentMatches.matches) {
+      final RouteBase route = match.route;
+      if (route is GoRoute && route.redirect != null) {
+        final String? redirectedLocation = route.redirect!(
+          GoRouterState(
+            configuration,
+            location: currentMatches.location.toString(),
+            subloc: match.subloc,
+            name: route.name,
+            path: route.path,
+            fullpath: match.fullpath,
+            extra: match.extra,
+            params: match.decodedParams,
+            queryParams: match.queryParams,
+            queryParametersAll: match.queryParametersAll,
+          ),
+        );
+        if (redirectedLocation != null) {
+          newMatchList = matcher.findMatch(redirectedLocation);
+          _addRedirect(redirects, newMatchList, prevMatchList.location,
+              configuration.redirectLimit);
+          break;
+        }
+      }
+    }
+    if (newMatchList != null) {
+      continue;
     }
 
-    final RouteBase topRoute = top.route;
-    assert(topRoute is GoRoute,
-        'Last RouteMatch should contain a GoRoute, but was ${topRoute.runtimeType}');
-    final GoRoute topGoRoute = topRoute as GoRoute;
-    final GoRouterRedirect? redirect = topGoRoute.redirect;
-    if (redirect == null) {
-      break;
-    }
-
-    final String? topRouteLocation = redirect(
-      GoRouterState(
-        configuration,
-        location: currentMatches.location.toString(),
-        subloc: top.subloc,
-        name: topGoRoute.name,
-        path: topGoRoute.path,
-        fullpath: top.fullpath,
-        extra: top.extra,
-        params: top.decodedParams,
-        queryParams: top.queryParams,
-        queryParametersAll: top.queryParametersAll,
-      ),
-    );
-
-    if (topRouteLocation == null) {
-      break;
-    }
-
-    final RouteMatchList newMatchList = matcher.findMatch(topRouteLocation);
-    _addRedirect(redirects, newMatchList, prevMatchList.location,
-        configuration.redirectLimit);
-    continue;
+    return currentMatches;
   }
-  return matches;
 }
 
 /// A configuration error detected while processing redirects.